Dev register form
This exercise allows developers to register with their names, technologies they know and how many years of experience they have working with each one of them.

By clicking in "Register", all the information is saved in an array, where you can easily access all the categories that was filled in the form.